+++ resources/views/admin/section/index.php
@@ -10,17 +10,17 @@ declare(strict_types=1);
  */
 
 use Hirtz\Cms\Modules\Admin\Data\SectionActiveDataProvider;
-use Hirtz\Cms\Modules\Admin\Widgets\Buttons\SectionCreateButton;
 use Hirtz\Cms\Modules\Admin\Widgets\Grids\SectionGridView;
 use Hirtz\Cms\Modules\Admin\Widgets\Navs\EntryHeader;
 use Hirtz\Cms\Modules\Admin\Widgets\Navs\EntrySubmenu;
+use Hirtz\Cms\Modules\Admin\Widgets\Navs\SectionActionDropdown;
 use Hirtz\Skeleton\Web\View;
 use Hirtz\Skeleton\Widgets\Grids\GridContainer;
 
 echo EntryHeader::make()
     ->model($provider->entry)
-    ->content(SectionCreateButton::make()
-        ->model($provider->entry));
+    ->content(SectionActionDropdown::make()
+        ->provider($provider));
 
 echo EntrySubmenu::make()
     ->model($provider->entry);